스타일 분리
